+++ src/components/RandomGenerator.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from "react"
-import { cryptoRandomUint32 } from "../../jacdac-ts/src/jdom/utils"
 import {
     deviceSpecificationFromFirmwareIdentifier,
     serviceSpecificationFromClassIdentifier,
@@ -20,6 +19,7 @@ import Alert from "./ui/Alert"
 import { Button } from "gatsby-theme-material-ui"
 import { NoSsr } from "@material-ui/core"
 import { useId } from "react-use-id-hook"
+import { cryptoRandomUint32 } from "../../jacdac-ts/src/jdom/random"
 
 function looksRandom(n: number) {
     const s = n.toString(16)
